refactor: Moved Dataverse details fetch to utils
